add estimation for swap in swap out depending on interest.
