+++ keeper/query_server.go
@@ -2,13 +2,8 @@ package keeper
 
 import (
 	"context"
-	"errors"
 	"fmt"
 
-	"cosmossdk.io/collections"
-	sdkmath "cosmossdk.io/math"
-
-	"github.com/provlabs/vault/interest"
 	"github.com/provlabs/vault/types"
 	"github.com/provlabs/vault/utils"
 	"google.golang.org/grpc/codes"
@@ -130,34 +125,6 @@ func (k queryServer) EstimateSwapIn(goCtx context.Context, req *types.QueryEstim
 		Height: ctx.BlockHeight(),
 		Time:   ctx.BlockTime().UTC(),
 	}, nil
-}
-
-func (k Keeper) EstimateVaultTotalAssets(ctx sdk.Context, vault *types.VaultAccount, totalAssets sdk.Coin) (sdkmath.Int, error) {
-	estimated := totalAssets.Amount
-
-	if vault.InterestRate == "" {
-		return estimated, nil
-	}
-
-	interestDetails, err := k.VaultInterestDetails.Get(ctx, vault.GetAddress())
-	if err != nil {
-		if !errors.Is(err, collections.ErrNotFound) {
-			return sdkmath.Int{}, fmt.Errorf("error getting interest details: %w", err)
-		}
-		return estimated, nil
-	}
-
-	duration := ctx.BlockTime().Unix() - interestDetails.PeriodStart
-	if duration <= 0 {
-		return estimated, nil
-	}
-
-	interestEarned, err := interest.CalculateInterestEarned(totalAssets, vault.InterestRate, duration)
-	if err != nil {
-		return sdkmath.Int{}, fmt.Errorf("error calculating interest: %w", err)
-	}
-
-	return estimated.Add(interestEarned), nil
 }
 
 // EstimateSwapOut estimates the amount of underlying assets that would be received for a given amount of shares.
